requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
async-lru==1.0.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.27.1
azure-identity==1.12.0
azure-keyvault-keys==4.7.0
azure-keyvault-secrets==4.6.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
cryptography==41.0.1
dependency-injector==4.41.0
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.21.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
isodate==0.6.1
joblib==1.2.0
llvmlite==0.40.1
Markdown==3.5.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
n2w==0.1.3
numba==0.57.1
numpy==1.24.1
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.8.0
packaging==23.1
pandas==1.5.3
pandasql==0.7.3
patsy==0.5.3
pluggy==1.2.0
portalocker==2.7.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.5
PyJWT==2.7.0
pytest==7.2.1
pytest-aiohttp==1.0.5
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.11.0
shap==0.42.1
six==1.16.0
slicer==0.0.7
SQLAlchemy==1.4.46
statsmodels==0.14.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
xgboost==1.7.3
yarl==1.9.2
